// src/common/interfaces/WebPartContext@1.15.2.ts
export class PageContextCopy152 {
    listItem: { 
      id: number;
    }
    list: { 
      title: string;
      id: string;
      serverRelativeUrl: string;
    }
    user: {
      displayName: string;
      loginName: string;
      email: string;
    }
    web: {
      title: string;
      serverRelativeUrl: string;
      id: string;
      absoluteUrl: string;
      language: string;
    }
    site: {
      // id: string;
      // serverRelativeUrl: string;
      absoluteUrl: string;
    }
}


export class WebPartContextCopy152 {

  pageContext: PageContextCopy152;

}